Fixes to allow using absl::int128 as Int128
Added tests:
* Basic cases for Int128 numeric columns
* Parsing Decimal128 values from string (basic and with overflow)

Known limitations:
* Overflow checks are broken for std::numeric_limits<Int128>::min() when there is no __int128.

+++ clickhouse/columns/date.cpp
@@ -134,7 +134,8 @@ void ColumnDateTime64::Append(const Int64& value) {
 //}
 
 Int64 ColumnDateTime64::At(size_t n) const {
-    return data_->At(n);
+    // make sure to use Absl's Int128 conversion
+    return static_cast<long long>(data_->At(n));
 }
 
 void ColumnDateTime64::Append(ColumnRef column) {
